reusable ui moved into :core:ui module
