+++ app/src/main/java/srimani7/apps/feedfly/ui/articles/LabelledArticleCard.kt
@@ -46,7 +46,12 @@ import kotlinx.coroutines.launch
 import srimani7.apps.feedfly.R
 import srimani7.apps.feedfly.core.design.FeedFlyTheme
 import srimani7.apps.feedfly.core.model.LabelledArticle
-import srimani7.apps.feedfly.util.fromHtml
+import srimani7.apps.feedfly.core.ui.ArticleDescription
+import srimani7.apps.feedfly.core.ui.ArticleImage
+import srimani7.apps.feedfly.core.ui.ArticleMediaHeader
+import srimani7.apps.feedfly.core.ui.ArticleTitle
+import srimani7.apps.feedfly.core.ui.ArticleViewScreen
+import srimani7.apps.feedfly.core.ui.util.fromHtml
 import srimani7.apps.rssparser.DateParser
 import java.util.Date
