+++ app/src/main/java/srimani7/apps/feedfly/ui/articles/PrivateArticleCard.kt
@@ -35,7 +35,12 @@ import kotlinx.coroutines.launch
 import srimani7.apps.feedfly.R
 import srimani7.apps.feedfly.core.design.FeedFlyTheme
 import srimani7.apps.feedfly.core.model.PrivateArticle
-import srimani7.apps.feedfly.util.fromHtml
+import srimani7.apps.feedfly.core.ui.ArticleDescription
+import srimani7.apps.feedfly.core.ui.ArticleImage
+import srimani7.apps.feedfly.core.ui.ArticleMediaHeader
+import srimani7.apps.feedfly.core.ui.ArticleTitle
+import srimani7.apps.feedfly.core.ui.ArticleViewScreen
+import srimani7.apps.feedfly.core.ui.util.fromHtml
 import srimani7.apps.rssparser.DateParser
 import java.util.Date
