+++ core/ui/build.gradle
@@ -27,6 +27,12 @@ android {
     kotlinOptions {
         jvmTarget = '1.8'
     }
+    buildFeatures {
+        compose true
+    }
+    composeOptions {
+        kotlinCompilerExtensionVersion libs.versions.composeCompiler.get()
+    }
 }
 
 dependencies {
@@ -36,4 +42,11 @@ dependencies {
     testImplementation libs.junit
     androidTestImplementation libs.androidx.junit
     androidTestImplementation libs.androidx.espresso.core
+
+    implementation platform(libs.androidx.compose.bom)
+
+    implementation libs.material3
+    implementation libs.coil.compose
+    implementation libs.androidx.browser
+
 }
