+++ core/ui/src/main/java/srimani7/apps/feedfly/core/ui/ArticleMediaHeader.kt
@@ -1,4 +1,4 @@
-package srimani7.apps.feedfly.ui.articles
+package srimani7.apps.feedfly.core.ui
 
 import androidx.compose.runtime.Composable
