+++ core/ui/src/main/java/srimani7/apps/feedfly/core/ui/ArticleTitle.kt
@@ -1,0 +1,18 @@
+package srimani7.apps.feedfly.core.ui
+
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.style.TextOverflow
+
+@Composable
+fun ArticleTitle(title: String?, modifier: Modifier = Modifier) {
+    if (!title.isNullOrBlank()) Text(
+        text = title,
+        style = MaterialTheme.typography.bodyLarge,
+        maxLines = 2,
+        overflow = TextOverflow.Ellipsis,
+        modifier = modifier
+    )
+}
