+++ core/ui/src/main/java/srimani7/apps/feedfly/core/ui/GroupsPicker.kt
@@ -1,6 +1,6 @@
 @file:OptIn(ExperimentalMaterial3Api::class, ExperimentalLayoutApi::class)
 
-package srimani7.apps.feedfly.ui
+package srimani7.apps.feedfly.core.ui
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
