+++ core/ui/src/main/res/drawable/open_in_browser_24px.xml
