+++ core/ui/src/main/res/drawable/share_24px.xml
